Migrate from f64 to rust_decimal for monetary precision
- Add rust_decimal and rust_decimal_macros dependencies
- Remove ordered-float dependency (no longer needed)
- Replace all f64 monetary amounts with Decimal across the entire stack:
  DataValue::Money, DataValue::Percentage, StatementTxn, TrialBalanceItem,
  LedgerEntryCommand, CreateJournalCommand, SetRateCommand, JournalEntry
- Update storage engine internals (LedgerStore, LedgerDay, RateStore) to use Decimal
- Update evaluator arithmetic to use Decimal operations
- Update accrual calculation to use Decimal with round_dp(2)
- Use MathematicalOps trait for Decimal exponentiation

+++ src/models/read.rs
@@ -1,5 +1,6 @@
 use std::{collections::BTreeMap, sync::Arc};
 
+use rust_decimal::Decimal;
 use time::Date;
 
 use super::DataValue;
@@ -7,6 +8,6 @@ use super::DataValue;
 pub struct JournalEntry {
     pub date: Date,
     pub description: Arc<str>,
-    pub amount: f64,
+    pub amount: Decimal,
     pub dimensions: BTreeMap<Arc<str>, Arc<DataValue>>,
 }
