+++ src/models/write.rs
@@ -1,5 +1,6 @@
 use std::{sync::Arc, collections::BTreeMap};
 
+use rust_decimal::Decimal;
 use time::Date;
 
 use super::DataValue;
@@ -9,23 +10,17 @@ use super::DataValue;
 pub struct CreateJournalCommand {
     pub date: Date,
     pub description: Arc<str>,    
-    pub amount: f64,
+    pub amount: Decimal,
     pub ledger_entries: Vec<LedgerEntryCommand>,
     pub dimensions: BTreeMap<Arc<str>, Arc<DataValue>>,
 }
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum LedgerEntryCommand {
-    Debit {account_id: Arc<str>, amount: f64},
-    Credit {account_id: Arc<str>, amount: f64},
+    Debit {account_id: Arc<str>, amount: Decimal},
+    Credit {account_id: Arc<str>, amount: Decimal},
 }
 
-
-// #[derive(Debug, Clone, PartialEq)]
-// pub struct CreateAccountCommand {
-//     pub id: Arc<str>,    
-    
-// }
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct CreateRateCommand {
@@ -37,5 +32,5 @@ pub struct CreateRateCommand {
 pub struct SetRateCommand {
     pub id: Arc<str>,
     pub date: Date,
-    pub rate: f64,
+    pub rate: Decimal,
 }
